fix: log handler
